requirements.txt
absl-py==1.3.0
alembic==1.8.1
asttokens==2.2.1
attrs==22.1.0
autopage==0.5.1
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cliff==4.1.0
cloudpickle==2.2.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
future==0.18.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.50.0
Gymnasium==0.26.3
gymnasium-notices==0.0.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.13.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
Minigrid==2.1.0
nest-asyncio==1.5.6
numpy==1.23.4
numpy-indexed==0.3.5
oauthlib==3.2.2
opencv-python==4.6.0.66
optuna==3.0.3
packaging==21.3
pandas==1.5.1
parso==0.8.3
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
prettytable==3.5.0
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.12.1
six==1.16.0
SQLAlchemy==1.4.44
stack-data==0.6.2
stevedore==4.1.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
threadpoolctl==3.1.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
